removed intmax_t. in c++11, intmax_t was defined to be 64bit even in 32bit platform. This causes severe performance problem in 32bit platform.

+++ cpp/lib_calvin/graph/graph_test.cc
@@ -6,7 +6,7 @@ using namespace lib_calvin_graph;
 void lib_calvin_graph::graphTest() {
 	std::cout << "--------------- Beginning graph test ----------------\n\n";
 
-	GraphTest<intmax_t, intmax_t> test(1000, 10000, 0);
+	GraphTest<int, int> test(1000, 10000, 0);
   test.insertionTest();
 	test.insertionTest2();
 	test.dfsTest();
